# backend-collect/doc/api.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Terminator111/COLLECT/5.0.0
info:
  description: 软工三项目接口文档
  version: 4.0.0
  title: COLLECT-API
...
paths:
...
  /admin/task/detail/{taskId}:
    get:
      tags:
        - admin
      summary: 管理员查看任务详细信息
      description: 管理员从任务列表中点击某一个任务，查看任务详细信息
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/TaskVO'
...
  /task/unfinished:
    get:
      tags:
        - task
      summary: 预览所有任务列表
      description: 发包方/众包工人在任务广场中，查看所有正在招募的任务列表（预览）
      parameters:
        - name: pageId
          in: query
          schema:
            type: integer
          required: true
        - name: userId
          in: query
          schema:
            type: integer
          required: true
        # - name: sortType
        #   in: query
        #   schema:
        #     type: integer
        #   required: true
        #   description: 0代表按发表时间倒序，1代表按最新回复时间倒序
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskViewVO"
  /task/detail/{taskId}:
    get:
      tags:
        - task
      summary: 查看任务详细信息
      description: 发包方/众包工人从任务列表中点击某一个任务，查看任务详细信息
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
        - name: userId
          in: query
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/TaskVO"
  /task/issue:
    post:
      tags:
        - task
      summary: 通过请求体发布任务
      description: 发包方通过PostDTO发布任务，taskId后端自行生成，发布任务后以VO格式返回，方便前端展示创建任务的详细信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskDTO"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/TaskVO"
  # 可以拆分成两个接口：选取、根据taskId获得taskVO
  /task/pick/{taskId}:
    post: # get？
      tags:
        - task
      summary: 选取任务
      description: 众包工人在任务详细页面可以选取任务，并获取任务相关资料，后端要存储选取的信息
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
        - name: userId
          in: query
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/TaskVO"
  /task/detail/getreport/{taskId}:
    get:
      tags:
        - task
      summary: 获取报告Id
      description: 当用户在任务详情页要去完善报告时，需要知道已经提交的报告Id
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
        - name: userId
          in: query
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: integer
                    description: reportId
...
  /task/taskRecommendFactor/set/{taskId}:
    post:
      tags:
        - task
      summary: 发包方设置此任务的推荐权重
      description: 发包方设置此任务的推荐权重
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/RecommendRuleFactorVO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecommendRuleFactorVO'
  /task/taskRecommendFactor/get/{taskId}:
    get:
      tags:
        - task
      summary: 发包方查看此任务的推荐权重
      description: 发包方查看此任务的推荐权重
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecommendRuleFactorVO'
  /task/taskRadar/current/{taskId}:
    get:
      tags:
        - task
      summary: 发包方查看此任务的当前工人的雷达图值
      description: 发包方查看此任务的当前工人的雷达图值，包括工人能力、活跃度、相关性、多样性
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                      $ref: '#/components/schemas/TaskRadarVO'
  /task/taskRadar/target/{taskId}:
    get:
      tags:
        - task
      summary: 发包方查看此任务的目标的雷达图值
      description: 发包方查看此任务的目标雷达图值，包括工人能力、活跃度、相关性、多样性
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                      $ref: '#/components/schemas/TaskRadarVO'
  /task/stop/{taskId}:
    post:
      tags:
        - task
      summary: 发包方手动停止招募
      description: 发包方手动停止招募
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseVO"
  /task/taskFlawDetection/{taskId}:
    get:
      tags:
        - task
      summary: 发包方查看此任务的全新缺陷发现曲线及预测的总缺陷数
      description: 发包方查看此任务的新缺陷发现曲线及预测的总缺陷数，注意，只计算每个缺陷树根节点的发现时间和数量，以一个缺陷树为单位。返回值是以天为单位，返回在xx天前一共发现了xx个缺陷
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/TaskFlawDetectionPredictionVO'
  /report/detail/{reportId}:
    get:
      tags:
        - report
      summary: 查看报告详细信息
      description: 发包方/众包工人从报告列表中点击某一个报告，查看报告详细信息
      parameters:
        - name: reportId
          in: path
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/ReportVO"
  /report/issue:
    post:
      tags:
        - report
      summary: 通过请求体提交报告
      description: 众包工人在线填写，并通过ReportDTO提交报告，reportId后端自行生成，发布报告后以VO格式返回，方便前端展示填写报告的详细信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/ReportVO'
  /report/toBeRefined:
    get:
      tags:
        - report
      summary: 获取报告中待修改的缺陷列表
      description: 在众包工人提交报告之后，系统会检测与该报告中的缺陷相似的缺陷，前端通过这个接口获取相似缺陷列表
      parameters:
        - name: reportId
          in: query
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TBAFlawVO'
...
components:
  schemas:
    ResponseVO:
      type: object
      description: api返回格式
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          type: object
...
    TaskVO:
      type: object
      description: 众测任务
      properties:
        id:
          type: integer
        userId:
          type: integer
          description: 发包方的id
        type:
          type: integer
          description: 0代表功能测试，1代表性能测试
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        workerNum:
          type: integer
        # others……
        pickedWorkerNum:
          type: integer
        exeFileName:
          type: string
        reqDocName:
          type: string
        state:
          type: integer
          description: 用来描述任务的状态。0表示未领取或者请求该任务详情的用户不是众包工人，1表示进行中，2表示已完成，3表示逾期未交，4表示待完善（缺陷没有完全被处理），5表示任务已停止招募
        createTime:
          type: string
          format: date-time
        tagGroups:
          type: array
          items:
            type: integer
    TaskViewVO:
      type: object
      description: 任务预览信息
      properties:
        id:
          type: integer
        userId:
          type: integer
          description: 发包方的id
        type:
          type: integer
          description: 0代表功能测试，1代表性能测试
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        workerNum:
          type: integer
        pickedWorkerNum:
          type: integer
        tagGroups:
          type: array
          items:
            type: integer
    TaskRadarVO:
      type: object
      description: 任务雷达图信息
      properties:
        ability:
          type: number
          description: 能力
        activation:
          type: number
          description: 活跃度
        revelance:
          type: number
          description: 相关性
        diversity:
          type: number
          description: 多样性
        avgTarget:
          type: number
          description: 平均目标值
    TaskFlawDetectionPredictionVO:
      type: object
      description: 包含对当前任务预测的所有可能flaw数量及已发现缺陷曲线列表
      properties:
        numPredicted:
          type: integer
          description: 对当前任务预测的所有可能不重复的flaw数量
        curves:
          type: array
          items:
            $ref: '#/components/schemas/TaskFlawDetectionCurveVO'
    TaskFlawDetectionCurveVO:
      type: object
      description: 
        在此日期及以前，发现了几个缺陷（以天为单位，只计算缺陷树的根节点）
      properties:
        time:
          type: string
          format: date-time
        num:
          type: integer
    TaskDTO:
      type: object
      description: 众测任务
      properties:
        userId:
          type: integer
          description: 发包方的id
        type:
          type: integer
          description: 0代表功能测试，1代表性能测试
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        workerNum:
          type: integer
        # others……
        # 以下 两个 实现方式 待定
        exeFileURL:
          type: string
        reqDocURL:
          type: string
        ## intention2: 要不要在发包方发布任务的时候让他给任务选定一组标签
        tagGroups:
          type: array
          items:
            type: integer
        recommendFactor:
          type: array
          items:
            $ref: '#/components/schemas/RecommendRuleFactorVO'
    ReportVO:
      type: object
      description: 测试报告
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
          description: 众包工人的id
        title:
          type: string
        createTime:
          type: string
          format: date-time
        ## change5: 我在这个类型中添加了一个score属性
        score:
          type: number
        flaws:
          type: array
          items:
            $ref: '#/components/schemas/FlawVO' ## change6: 我修改了flawVO，这里引用了flawVO
    ReportDTO:
      type: object
      description: 测试报告传输
      properties:
        taskId:
          type: integer
        userId:
          type: integer
          description: 众包工人的id
        title:
          type: string
        flaws:
          type: array
          items:
            ## change1: 这里原来使用的是FlawVO，我将它改为FlawDTO
            $ref: '#/components/schemas/FlawDTO'
...
    FlawDTO:
      type: object
      description: 缺陷
      properties:
        pictureURLs:
          type: array
          items:
            type: string
        description:
          type: string
        stepDes:
          type: string
        deviceInfo:
          type: string
    ## change3: 我添加了一个flawVO类型，相比原来的flawVO，这个类型中多出了score和forkedId属性
...
    FlawVO:
      type: object
      description: 缺陷
      properties:
        id:
          type: integer
        reportId:
          type: integer
        pictureURLs:
          type: array
          items:
            type: string
        description:
          type: string
        stepDes:
          type: string
        deviceInfo:
          type: string
        score:
          type: number
        scoreNum:
          type: integer
          description: 评分人数
        state:
          type: boolean
          description: 该缺陷是否已经被处理
        appendContent:
          type: string
          description: 对已提交的报告的补充内容
...
    SimilarFlawVO:
      type: object
      description: 相似缺陷
      properties:
        similarity:
          type: number
          maximum: 1.0
          minimum: 0.0
        flaw:
          $ref: '#/components/schemas/FlawVO'
    TBAFlawVO:
      type: object
      description: 有待完善的缺陷
      properties:
        tbaFlaw:
          $ref: '#/components/schemas/FlawVO'
        similarFlaws:
          type: array
          items:
            $ref: '#/components/schemas/SimilarFlawVO'
...
    RecommendRuleFactorVO:
      type: object
      description: 推荐优先级影响因素
      properties:
        name:
          type: string
        weight:
          type: number
          maximum: 1.0
          minimum: 0.0
...